requirements.txt
langgraph
langchain[openai]
langchain 
langchain-openai 
python-dotenv
requests
pydantic
Flask
flask-cors